Refactoring of commander and receiver
